claude-code-memory/requirements.txt
mem0ai>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
pydantic>=2.5.0
qdrant-client>=1.7.0
ollama>=0.1.0